// src/shared/models/minecraft-server-properties.model.ts
export interface IMinecraftServerProps {
  'spawn-protection': string;
  'max-tick-time': string;
  'query.port': string;
  'generator-settings': string;
  'force-gamemode': string;
  'allow-nether': string;
  'enforce-whitelist': string;
  gamemode: string;
  'broadcast-console-to-ops': string;
  'enable-query': string;
  'player-idle-timeout': string;
  difficulty: string;
  'spawn-monsters': string;
  'broadcast-rcon-to-ops': string;
  'op-permission-level': string;
  pvp: string;
  'snooper-enabled': string;
  'level-type': string;
  hardcore: string;
  'enable-command-block': string;
  'max-players': string;
  'network-compression-threshold': string;
  'resource-pack-sha1': string;
  'max-world-size': string;
  'function-permission-level': string;
  'rcon.port': string;
  'server-port': string;
  'texture-pack': string;
  'server-ip': string;
  'spawn-npcs': string;
  'allow-flight': string;
  'level-name': string;
  'view-distance': string;
  'resource-pack': string;
  'spawn-animals': string;
  'white-list': string;
  'rcon.password': string;
  'generate-structures': string;
  'online-mode': string;
  'max-build-height': string;
  'level-seed': string;
  'use-native-transport': string;
  'prevent-proxy-connections': string;
  motd: string;
  'enable-rcon': string;
}

export interface IEditableMinecraftServerProps {
  'spawn-protection': string;
  // 'max-tick-time': string;
  // 'query.port': string;
  // 'generator-settings': string;
  'force-gamemode': string;
  'allow-nether': string;
  // 'enforce-whitelist': string;
  gamemode: string;
  'broadcast-console-to-ops': string;
  // 'enable-query': string;
  'player-idle-timeout': string;
  difficulty: string;
  'spawn-monsters': string;
  'broadcast-rcon-to-ops': string;
  'op-permission-level': string;
  pvp: string;
  // 'snooper-enabled': string;
  'level-type': string;
  hardcore: string;
  'enable-command-block': string;
  'max-players': string;
  'network-compression-threshold': string;
  // 'resource-pack-sha1': string;
  'max-world-size': string;
  'function-permission-level': string;
  // 'rcon.port': string;
  // 'server-port': string;
  // 'texture-pack': string;
  // 'server-ip': string;
  'spawn-npcs': string;
  'allow-flight': string;
  'level-name': string;
  'view-distance': string;
  // 'resource-pack': string;
  'spawn-animals': string;
  'white-list': string;
  // 'rcon.password': string;
  'generate-structures': string;
  'online-mode': string;
  'max-build-height': string;
  'level-seed': string;
  // 'use-native-transport': string;
  // 'prevent-proxy-connections': string;
  motd: string;
  // 'enable-rcon': string;
}